Fix lib unit tests
